chore: Add annotations for data binding
